chore(reorg, sequencer): moved bt principic things to its own folder
